Move client generation to a submodule
